[PATCH] dup_mmap() double memory accounting
dup_mmap() unnecessarily tries to account for memory of the vma's it has
created if it fails in the middle.

However, that's pointless (and wrong), since the exit_mmap() path called
through mmput() will do so anyway in the failure path.

Just remove the bogus un-accounting code.

BKrev: 40c73851QKveTlOiknnZIEo4yrj7yg