[PATCH] Fix x86-64 VIA systems with IOMMU debug
VIA chipsets are broken and don't work with IOMMU enabled. For some mysterious
reason (the IOMMU logic is actually in the CPU) they manage to corrupt all
data going through the aperture to PCI devices.

There was a workaround previously that enabled the slower softmmu when VIA is
detected on the normal IOMMU path (when there is more than 3GB of memory). But
CONFIG_IOMMU_DEBUG took a short cut that was not handled, which ended up with
VIA systems not booting when this option is enabled.

This patch enables the workaround with CONFIG_IOMMU_DEBUG/iommu=force too

BKrev: 40ce055bJjqvIL_4yveL_FTRrjYg3Q