[PATCH] ahc1542 !CONFIG_MCA build fix
The old 1542 scsi driver is both ISA and MCA.  The MCA portions are disabled
when !CONFIG_MCA through the typical wrapper scheme (a la pci.h and
!CONFIG_PCI).  However...  the driver unconditionally includes linux/mca.h
which in turn unconditionally includes asm/mca.h.

This breaks drivers on platforms with ISA but not MCA, like alpha.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 40d30399yaNCfEt8BTXpfknskYHZvA