Merge bk://bk.arm.linux.org.uk/linux-2.6-rmk into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/06/18 13:33:48+01:00 rmk
[ARM] Add support code for ARM hardware vector floating point

This cset adds the code to handle the hardware vector floating point
unit found on some ARM926 and later CPUs.  The hardware provides
an implementation for the common cases, and bounces exceptions for
other cases, which have to be handled in software, and signalling
SIGFPE as appropriate.

2004/06/15 16:33:27+01:00 rmk
[ARM] pxafb doesn't need to include asm/mach-types.h

2004/06/15 15:50:19+01:00 elf
[ARM PATCH] 1928/1: lh7a40x #7 Changes to memory model to support contiguous SDRAM

Patch from Marc Singer

[if's replaced by ifdef's.]

While the BLOB boot loader doesn't support it, there are two others,
UBOOT and Logic's LOLO, that will initialize the SDRAM controller such
that the memory appears as a contiguous region.  This layout to be
required for WinCE and, thus, is the default for these loaders.  The
memory organization had been optimized for discontiguous blocks.  Now,
it supports a default mode that is compatible with all of the
bootloaders as well as options to optimize for either contiguous or
discontiguous models.

Withing, there is also a slight change to the way that the DEBUG_LL
macros select the UART address.  Now it uses immediate constants
exclusively.


2004/06/15 15:46:21+01:00 icampbell
[ARM PATCH] 1926/3:  PXAFB cleanups and fixes

Patch from Ian Campbell

Sorry, 1926/2 was missing a brace. Fixed here.

Previous commentry:
I have corrected the grayscale conversion to correctly use the upper
byte rather than the lower one. As requested I have also left the
actual grayscale conversion where it was this time and just dealt with
the palette differences in pxafb_setpalettereg.

fb_alloc_cmap is only called when bits_per_pixel is < 16, otherwise
any existing cmap is freed.

I also fixed up all the spaces and brackets issues (not just the one
you raised previously) this introduces some noise in the patch, if
you'd like the two patches separately please ask.

2004/06/15 14:52:36+01:00 rmk
[ARM] Add clock API

Since peripheral clocks are be derived in SoC or even platform
specific ways, we need a way to keep this information out of the
drivers.  AMBA Primecells are defined in terms of functional units
where clocks are derived from external sources, and it is up to the
SoC designer to determine where to derive those clocks from.

Therefore, we provide a very basic API which allows platforms to
provide this information in a generic manner to their peripherals.
This framework also allows peripherals to shut down unused clock
sources when they're not in use.

BKrev: 40d3183e4d9k6kS1MxinsCVU3oa5xA