Merge davem@nuts.davemloft.net:/disk1/BK/net-2.6 into kernel.bkbits.net:/home/davem/net-2.6
2004/06/18 14:14:17-07:00 akpm
[NET]: Fix eql.c failed dev_get_by_name() return value check.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/06/18 14:12:29-07:00 yoshfuji
[IPV6]: Fix autoconf description in ip-sysctl.txt.

2004/06/18 14:11:36-07:00 wli
[IRDA]: Remove usage of isa_virt_to_bus()

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/06/18 14:09:17-07:00 dlstevens
[IPV6]: Handle user asking for any device in mcast calls.

2004/06/18 13:54:52-07:00 shemminger
[PKT_SCHED]: Add loss option to network delay scheduler.

This enhances the network simulation scheduler to do simple random loss.

The loss parameter is a simple 32 bit value such that 0 means no loss, and
0xffffffff is always drop.  I have a new version of the tc command which takes
care of conversion from percent to this value.

Same patch for 2.4 and 2.6

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/06/18 13:51:42-07:00 davem
[PKT_SCHED]: Do not check netif_queue_stopped() in dequeue ops, races with driver.

Based upon a patch from Stephen Hemminger.
Signed-off-by: David S. Miller <davem@redhat.com>

2004/06/18 13:40:49-07:00 shemminger
[PKT_SCHED]: Delay scheduler should retry if requeue fails.

If delay scheduler decides not to send the packet right away, it requeues
it.  If the requeue fails, it should go and look again rather than waking
up prematurely.

Same patch should apply to both 2.6 and 2.4

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/06/18 13:18:34-07:00 shemminger
[PKT_SCHED]: Delay scheduler enqueue always succeeds.

If underlying fifo enqueue fails, return the status not 0.
Same patch should apply to both 2.6 and 2.4

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/06/18 13:15:58-07:00 shemminger
[BRIDGE]: Kill sysfs hotplug avoidance hacks.

The whole effort to avoid hotplug was misguided.  If it is really a problem
(which it doesn't appear to be) then it can more easily be addressed by smarter
hotplug scripts in user space.

This patch gets rid of the whole subsystem hack for bridge kobjects.

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

BKrev: 40d36453_dm8_es90y48giVNc8yRYw