[PATCH] Add ovcamchip driver
This patch adds a new driver for the OmniVision OV6xx0 and OV7xx0 series of
CMOS image sensors.  It is currently used by the w9968cf USB webcam driver,
which is already in mainline 2.6.  Up until now it had to be compiled
outside the kernel tree, which is clearly suboptimal.

It is also used by version 2 of the ov511 USB webcam driver, which will be
merged in the near future.  That will reduce some code duplication, since
the existing ov511 has much of this code built-in.

This was previously submitted to Linux-USB-Devel, and I have fixed the
concerns that came up at that time.

Developer's Certificate of Origin 1.0

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
        have the right to submit it under the open source license
        indicated in the file; or

(b) The contribution is based upon previous work that, to the best
        of my knowledge, is covered under an appropriate open source
        license and I have the right under that license to submit that
        work with modifications, whether created in whole or in part
        by me, under the same open source license (unless I am
        by me, under the same open source license (unless I am
        permitted to submit under a different license), as indicated
        in the file; or

(c) The contribution was provided directly to me by some other
        person who certified (a), (b) or (c) and I have not modified
        it.

Signed-off-by: Mark McClelland <mark@alpha.dyndns.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 40d63fbfy1hUEohbVEIwyF9YS84KMw